app/requirements.txt
Flask==2.3.3
torch==2.1.0
Pillow==9.5.0